Remove datetime from the workerInvocationsAdaptive query
We currently have `datetime` and `status` as dimensions for the
CloudFlare Worker invocation metrics GraphQL query.

This causes CloudFlare workers to output a number of datapoints for
each unique tuple of dimensions, across several `datetime`s.

For worker requests and errors this is fine as we `.Add(...)` to those
counters, but for quantiles this causes us to only export the last
datapoints quantiles, which may be an outlier.

---

Removing the `datetime` for the query provides pre-aggregated
datapoints for us to process.

Further to this, `datetime` is not currently being captured in the
response data structure.