Export worker invocation status on metrics
This change adds `status` to the CloudFlare worker metrics being
exported.

This fixes an issue where, similar to `datetime` previously, we are
getting multiple metric series from the combinations of unique
dimensions.

This resulted in only the last datapoint series being exported from
the `cloudflare-exporter` for the CloudFlare workers.

Additionally this provides more information as the `status` can be
used to differentiate the internal errors seen in CloudFlare workers.